fix đống user

+++ src/main/java/com/lmsu/controller/staff/DeleteStaffServlet.java
@@ -1,21 +1,22 @@
-package com.lmsu.controller.member;
+package com.lmsu.controller.staff;
 
+import com.lmsu.controller.member.DeleteUserServlet;
 import com.lmsu.users.UserDAO;
 import com.lmsu.users.UserDTO;
 import org.apache.log4j.Logger;
 
 import javax.naming.NamingException;
-import javax.servlet.http.*;
-import javax.servlet.annotation.WebServlet;
 import javax.servlet.*;
+import javax.servlet.http.*;
+import javax.servlet.annotation.*;
 import java.io.IOException;
 import java.sql.SQLException;
 
-@WebServlet(name = "DeleteUserServlet", value = "/DeleteUserServlet")
-public class DeleteUserServlet extends HttpServlet {
+@WebServlet(name = "DeleteStaffServlet", value = "/DeleteStaffServlet")
+public class DeleteStaffServlet extends HttpServlet {
 
-    private static final String SHOW_MEMBER_CONTROLLER = "ShowMemberServlet";
-    private static final String SEARCH_USER_CONTROLLER = "SearchUserServlet";
+    private static final String SHOW_STAFF_CONTROLLER = "ShowStaffServlet";
+    private static final String SEARCH_STAFF_CONTROLLER = "SearchStaffServlet";
     static final Logger LOGGER = Logger.getLogger(DeleteUserServlet.class);
 
     protected void processRequest(HttpServletRequest request, HttpServletResponse response)
@@ -23,26 +24,37 @@ public class DeleteUserServlet extends HttpServlet {
         response.setContentType("text/html;charset=UTF-8");
         request.setCharacterEncoding("UTF-8");
 
-        String url = SHOW_MEMBER_CONTROLLER;
+        String url = SHOW_STAFF_CONTROLLER;
         String userID = request.getParameter("userPk");
         String searchValue = request.getParameter("txtSearchValue");
 
         try {
             UserDAO dao = new UserDAO();
+            HttpSession session = request.getSession();
+            UserDTO loginUser = (UserDTO) session.getAttribute("LOGIN_USER");
+            String loginUserID = "";
 
-            boolean result = dao.deleteUser(userID);
-
-            if (result) {
-                if (searchValue == null || searchValue.trim().isEmpty()) {
-                    url = SHOW_MEMBER_CONTROLLER;
-                } else {
-                    url = SEARCH_USER_CONTROLLER;
-                }
+            if (loginUser != null) {
+                loginUserID = loginUser.getId();
             }
-        } catch (SQLException ex) {
+
+            if (!loginUserID.equals(userID)){
+                boolean result = dao.deleteUser(userID);
+
+                if (result) {
+                    if (searchValue == null || searchValue.trim().isEmpty()) {
+                        url = SHOW_STAFF_CONTROLLER;
+                    } else {
+                        url = SEARCH_STAFF_CONTROLLER;
+                    }
+                }
+            } else {
+                request.setAttribute("LOGGING_IN_USER", "You're now logging into this account. Cannot delete it!");
+            }
+        } catch (SQLException ex){
             LOGGER.error(ex.getMessage());
             log("DeleteUserServlet _ SQL: " + ex.getMessage());
-        } catch (NamingException ex) {
+        } catch (NamingException ex){
             LOGGER.error(ex.getMessage());
             log("DeleteUserServlet _ Naming: " + ex.getMessage());
         } finally {
@@ -53,11 +65,11 @@ public class DeleteUserServlet extends HttpServlet {
 
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        processRequest(request, response);
+
     }
 
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        processRequest(request, response);
+
     }
 }
